creating ChatSettingsParsing class for chat settings parsing
